+++ Src/FluentAssertions/Formatting/Formatter.cs
@@ -20,6 +20,7 @@ public static class Formatter
 
     private static readonly List<IValueFormatter> DefaultFormatters = new()
     {
+        new PassthroughValueFormatter(),
         new XmlReaderValueFormatter(),
         new XmlNodeFormatter(),
         new AttributeBasedFormatter(),